Implemented Session for Post and Package

+++ frontend/src/app/api/post/route.js
@@ -41,49 +41,81 @@ export async function POST(req) {
 
     if (!title || !description || !category || !trainerid) return NextResponse.json({ "code": 400, "message": "Missing required data" })
 
-    const res = await fetch(`${DATA_SOURCE_URL}/post/create`, {
-        method: 'POST',
-        headers: {
-            "Content-Type": "application/json",
-            'API-Key': process.env.DATA_API_KEY
-        },
-        body: JSON.stringify({
-            title: title,
-            description: description,
-            category: category,
-            trainerid: trainerid,
-            image: image
-        })
-    })
+    try {
+        const headersInstance = headers()
+        const authHeader = headersInstance.get('Authorization')
 
-    const result = await res.json()
-    return NextResponse.json(result)
+        const token = authHeader.split(' ')[1]
+        const decoded = jwt.verify(token, process.env.JWT_SECRET)
+
+        if (!decoded) {
+            return NextResponse.json({ "code": 400, "message": "Expired" })
+        } else if (decoded.exp < Math.floor(Date.now() / 1000)) {
+            return NextResponse.json({ "code": 400, "message": "Expired" })
+        } else {
+            const res = await fetch(`${DATA_SOURCE_URL}/post/create`, {
+                method: 'POST',
+                headers: {
+                    "Content-Type": "application/json",
+                    'API-Key': process.env.DATA_API_KEY
+                },
+                body: JSON.stringify({
+                    title: title,
+                    description: description,
+                    category: category,
+                    trainerid: trainerid,
+                    image: image
+                })
+            })
+
+            const result = await res.json()
+            return NextResponse.json(result)
+        }
+    } catch (error) {
+        console.error('Token verification failed', error)
+        return NextResponse.json({ "code": 400, "message": "Unauthorized" })
+    }
 }
 
 export async function PUT(req) {
     const { title, description, category, image, trainerid, postid } = await req.json()
 
-    console.log(title, description, category, image, trainerid, postid);
-
     if (!title || !description || !category || !trainerid) return NextResponse.json({ "code": 400, "message": "Missing required data" })
 
-    const res = await fetch(`${DATA_SOURCE_URL}/post/${postid}`, {
-        method: 'PUT',
-        headers: {
-            "Content-Type": "application/json",
-            'API-Key': process.env.DATA_API_KEY
-        },
-        body: JSON.stringify({
-            title: title,
-            description: description,
-            category: category,
-            image: image,
-            trainerid: trainerid
-        })
-    })
+    try {
+        const headersInstance = headers()
+        const authHeader = headersInstance.get('Authorization')
 
-    const result = await res.json()
-    return NextResponse.json(result)
+        const token = authHeader.split(' ')[1]
+        const decoded = jwt.verify(token, process.env.JWT_SECRET)
+
+        if (!decoded) {
+            return NextResponse.json({ "code": 400, "message": "Expired" })
+        } else if (decoded.exp < Math.floor(Date.now() / 1000)) {
+            return NextResponse.json({ "code": 400, "message": "Expired" })
+        } else {
+            const res = await fetch(`${DATA_SOURCE_URL}/post/${postid}`, {
+                method: 'PUT',
+                headers: {
+                    "Content-Type": "application/json",
+                    'API-Key': process.env.DATA_API_KEY
+                },
+                body: JSON.stringify({
+                    title: title,
+                    description: description,
+                    category: category,
+                    image: image,
+                    trainerid: trainerid
+                })
+            })
+
+            const result = await res.json()
+            return NextResponse.json(result)
+        }
+    } catch (error) {
+        console.error('Token verification failed', error)
+        return NextResponse.json({ "code": 400, "message": "Unauthorized" })
+    }
 }
 
 export async function DELETE(req) {
@@ -91,17 +123,34 @@ export async function DELETE(req) {
 
     if (!postid) return NextResponse.json({ "code": 400, "message": "Failed to delete post" })
 
-    const res = await fetch(`${DATA_SOURCE_URL}/post/${postid}`, {
-        method: 'DELETE',
-        headers: {
-            "Content-Type": "application/json",
-            'API-Key': process.env.DATA_API_KEY
-        },
-        body: JSON.stringify({
-            postid: postid
-        })
-    })
+    try {
+        const headersInstance = headers()
+        const authHeader = headersInstance.get('Authorization')
 
-    const result = await res.json()
-    return NextResponse.json(result)
+        const token = authHeader.split(' ')[1]
+        const decoded = jwt.verify(token, process.env.JWT_SECRET)
+
+        if (!decoded) {
+            return NextResponse.json({ "code": 400, "message": "Expired" })
+        } else if (decoded.exp < Math.floor(Date.now() / 1000)) {
+            return NextResponse.json({ "code": 400, "message": "Expired" })
+        } else {
+            const res = await fetch(`${DATA_SOURCE_URL}/post/${postid}`, {
+                method: 'DELETE',
+                headers: {
+                    "Content-Type": "application/json",
+                    'API-Key': process.env.DATA_API_KEY
+                },
+                body: JSON.stringify({
+                    postid: postid
+                })
+            })
+
+            const result = await res.json()
+            return NextResponse.json(result)
+        }
+    } catch (error) {
+        console.error('Token verification failed', error)
+        return NextResponse.json({ "code": 400, "message": "Unauthorized" })
+    }
 }
